feature: add the ability to display items in the cart
It also includes the creation of new structures, CartSection and CartItemCellViewModel, to represent cart items in a collection view. The CartPresenter is updated with a new cartUseCase to fetch cart items and update the view accordingly. The CartConfigurator is also modified to ensure that the necessary gateways and use cases are set up for the cart.