add(CartRootView.swift, FavoriteRootView.swift): display the background view when the collection is empty
In the CartRootView.swift and FavoriteRootView.swift files, the appearance has been updated to display a message and image when there are no items in the cart or favorites list. A background view has been added that is hidden when the collection view has content. An observer has been added to the collection view's contentSize property to toggle the visibility of the background view